[feat] Add data live-update functionality to Board (#2796)
