+++ src/aimcore/web/ui/src/pages/App/App.d.ts
@@ -1,9 +1,16 @@
 import { BoardData } from 'modules/core/api/boardsApi';
 
-export interface AppStructureProps {
+export interface AppSidebarProps {
   boards: string[];
   editMode: boolean;
 }
+
+export type AppSidebarNode = {
+  title: string | React.ReactNode;
+  key: string;
+  value: string;
+  children?: AppSidebarNode[];
+};
 
 export interface AppWrapperProps {
   boardPath: string;
